updated metadata rendering to render individual links when grouped together
