+++ frontend/components/Viewing/MetadataInfo.js
@@ -12,27 +12,27 @@ export default function MetadataInfo({ title, link, label, icon, specific }) {
   //If the metadata doesn't contain the title nothing should be rendered.
   if (!title) return null;
 
-  return (
-    <Link href={link}>
-      <a target="_blank">
-        <div className={styles.info}>
-          <div className={specific ? styles.infogeneric : styles.infoheader}>
-            <div className={styles.infoiconcontainer}>
-              <FontAwesomeIcon
-                icon={icon}
-                size="1x"
-                className={styles.infoicon}
-              />
-            </div>
-            <div className={styles.infolabel}>{label}</div>
-          </div>
-          <div
-            className={specific ? styles.infotitlegeneric : styles.infotitle}
-          >
-            {title}
-          </div>
+  const renderedSection = (
+    <div className={styles.info}>
+      <div className={specific ? styles.infogeneric : styles.infoheader}>
+        <div className={styles.infoiconcontainer}>
+          <FontAwesomeIcon icon={icon} size="1x" className={styles.infoicon} />
         </div>
-      </a>
-    </Link>
+        <div className={styles.infolabel}>{label}</div>
+      </div>
+      <div className={specific ? styles.infotitlegeneric : styles.infotitle}>
+        {title}
+      </div>
+    </div>
   );
+
+  console.log(link);
+
+  if (link)
+    return (
+      <Link href={link}>
+        <a target="_blank">{renderedSection}</a>
+      </Link>
+    );
+  else return renderedSection;
 }
